feat(networkshim): consolidate code & export constants and util functions
