+++ packages/cypress-commands/src/helper/index.js
@@ -4,4 +4,3 @@ export {
     getDhis2BaseUrl,
     setDhis2BaseUrlToLocalStorage,
 } from './dhis2BaseUrl.js'
-export { isLiveMode, isCaptureMode, isStubMode } from './networkMode.js'
