--- packages/cypress-commands/src/helper/networkMode.js
@@ -1,6 +1,0 @@
-export const isLiveMode = () =>
-    !Cypress.env('networkMode') || Cypress.env('networkMode') === 'live'
-
-export const isCaptureMode = () => Cypress.env('networkMode') === 'capture'
-
-export const isStubMode = () => Cypress.env('networkMode') === 'stub'
