+++ packages/cypress-commands/src/setups/enableAutoLogin/index.js
@@ -1,0 +1,1 @@
+export { enableAutoLogin } from './enableAutoLogin.js'
